+++ lurk_macro/tests/macro_test.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod test {
     use blstrs::Scalar as Fr;
-    use lurk::{eval::IO, store::Store};
+    use lurk::{eval::lang::Lang, eval::IO, store::Store};
     use lurk_macro::{let_store, lurk};
 
     #[test]
@@ -62,6 +62,8 @@ mod test {
                           ((exp 5) 3)))
         .unwrap();
 
+        let lang = Lang::new();
+
         let (
             IO {
                 expr: result_expr,
@@ -70,7 +72,7 @@ mod test {
             },
             iterations,
             _emitted,
-        ) = Evaluator::new(expr, empty_sym_env(&s_), &mut s_, limit)
+        ) = Evaluator::new(expr, empty_sym_env(&s_), &mut s_, limit, &lang)
             .eval()
             .unwrap();
 
@@ -85,6 +87,7 @@ mod test {
         let limit = 20;
         let val = s_.num(123);
         let expr = lurk!(((lambda (x) x) 123)).unwrap();
+        let lang = Lang::new();
 
         let (
             IO {
@@ -94,7 +97,7 @@ mod test {
             },
             iterations,
             _emitted,
-        ) = Evaluator::new(expr, empty_sym_env(&s_), &mut s_, limit)
+        ) = Evaluator::new(expr, empty_sym_env(&s_), &mut s_, limit, &lang)
             .eval()
             .unwrap();
