+++ src/coprocessor.rs
@@ -26,3 +26,44 @@ pub trait Coprocessor<F: LurkField>: Debug + Sync {
 
     fn simple_evaluate(&self, s: &mut Store<F>, args: &[Ptr<F>]) -> Ptr<F>;
 }
+
+#[cfg(test)]
+pub(crate) mod test {
+    use super::*;
+    use std::marker::PhantomData;
+
+    /// A dumb Coprocessor for testing.
+    #[derive(Debug)]
+    pub(crate) struct DumbCoprocessor<F: LurkField> {
+        pub(crate) _p: PhantomData<F>,
+    }
+
+    impl<F: LurkField> Coprocessor<F> for DumbCoprocessor<F> {
+        /// Dumb Coprocessor takes two arguments.
+        fn arity(&self) -> usize {
+            2
+        }
+
+        /// It squares the first arg and adds it to the second.
+        fn simple_evaluate(&self, s: &mut Store<F>, args: &[Ptr<F>]) -> Ptr<F> {
+            let a = args[0];
+            let b = args[1];
+
+            let a_num = s.fetch_num(&a).unwrap();
+            let b_num = s.fetch_num(&b).unwrap();
+            let mut result = *a_num;
+            result *= *a_num;
+            result += *b_num;
+
+            s.intern_num(result)
+        }
+    }
+
+    impl<F: LurkField> DumbCoprocessor<F> {
+        pub(crate) fn new() -> Self {
+            Self {
+                _p: Default::default(),
+            }
+        }
+    }
+}
